AMLS-5522 | Further address package refactor, removed commented out code and duplicated files
