+++ app/controllers/responsiblepeople/address/CurrentAddressDateOfChangeController.scala
@@ -31,11 +31,9 @@ import utils.{AuthAction, DateOfChangeHelper, RepeatingSection}
 
 import scala.concurrent.Future
 
-class CurrentAddressDateOfChangeController @Inject()(
-                                                      val dataCacheConnector: DataCacheConnector,
-                                                      authAction: AuthAction,
-                                                      statusService: StatusService
-                                                    ) extends RepeatingSection with DefaultBaseController with DateOfChangeHelper with FormHelpers {
+class CurrentAddressDateOfChangeController @Inject()(val dataCacheConnector: DataCacheConnector,
+                                                     authAction: AuthAction,
+                                                     statusService: StatusService) extends RepeatingSection with DefaultBaseController with DateOfChangeHelper with FormHelpers {
 
   def get(index: Int, edit: Boolean) = authAction {
     implicit request =>
