+++ app/controllers/responsiblepeople/address/CurrentAddressNonUKController.scala
@@ -23,7 +23,7 @@ import forms.{EmptyForm, Form2, InvalidForm, ValidForm}
 import models.responsiblepeople._
 import services.{AutoCompleteService, StatusService}
 import uk.gov.hmrc.play.audit.http.connector.AuditConnector
-import utils.{AuthAction, ControllerHelper, DateOfChangeHelper, RepeatingSection}
+import utils.{AuthAction, ControllerHelper, DateOfChangeHelper}
 import views.html.responsiblepeople.address.current_address_NonUK
 
 import scala.concurrent.Future
