+++ app/controllers/responsiblepeople/address/NewHomeAddressNonUKController.scala
@@ -23,7 +23,7 @@ import javax.inject.{Inject, Singleton}
 import models.responsiblepeople.NewHomeAddress._
 import models.responsiblepeople._
 import services.AutoCompleteService
-import utils.{AuthAction, ControllerHelper, RepeatingSection}
+import utils.{AuthAction, ControllerHelper}
 import views.html.responsiblepeople.address.new_home_address_NonUK
 
 import scala.concurrent.Future
