+++ app/views/responsiblepeople/address/new_home_date_of_change.scala.html
@@ -29,7 +29,7 @@
   heading = header
 ) {
 
-  @form(f, controllers.responsiblepeople.routes.NewHomeAddressDateOfChangeController.post(index)) {
+  @form(f, controllers.responsiblepeople.address.routes.NewHomeAddressDateOfChangeController.post(index)) {
 
     @date(
       f = f,
