+++ app/views/responsiblepeople/detailed_answers.scala.html
@@ -89,7 +89,7 @@
         @if(showHide){
         <div class="panel-indent panel-indent--gutter panel-border-wide info">
             <p>
-                <a href="@controllers.responsiblepeople.routes.NewHomeAddressDateOfChangeController.get(idx)">
+                <a href="@controllers.responsiblepeople.address.routes.NewHomeAddressDateOfChangeController.get(idx)">
                     @Messages("responsiblepeople.detailed_answer.tell.us.moved", personName)
                 </a>
             </p>
@@ -276,7 +276,7 @@
                     question = Messages("responsiblepeople.detailed_answers.address.UK", personName),
                     editUrl = showHide match {
                         case true => {
-                            controllers.responsiblepeople.routes.MovedAddressController.get(idx).toString()
+                            controllers.responsiblepeople.address.routes.MovedAddressController.get(idx).toString()
                         }
                         case _ => {
                             controllers.responsiblepeople.address.routes.CurrentAddressController.get(idx, true, flow).toString()
@@ -292,7 +292,7 @@
                     question = Messages("responsiblepeople.detailed_answers.address", personName),
                     editUrl =  (a.personAddress, showHide)match {
                             case (_, true) => {
-                                controllers.responsiblepeople.routes.MovedAddressController.get(idx).toString
+                                controllers.responsiblepeople.address.routes.MovedAddressController.get(idx).toString
                             }
                             case (_:PersonAddressNonUK, false) => {
                                 controllers.responsiblepeople.address.routes.CurrentAddressNonUKController.get(idx, true, flow).toString()
