+++ test/controllers/responsiblepeople/address/CurrentAddressControllerNonUKSpec.scala
@@ -34,8 +34,6 @@ import org.jsoup.select.Elements
 import org.mockito.ArgumentCaptor
 import org.mockito.Matchers.{any, eq => eqTo}
 import org.mockito.Mockito._
-import org.scalatest.BeforeAndAfter
-import org.scalatest.mock.MockitoSugar
 import play.api.i18n.Messages
 import play.api.test.Helpers._
 import services.{AutoCompleteService, StatusService}
