+++ test/views/responsiblepeople/address/new_home_date_of_changeSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package views.responsiblepeople
+package views.responsiblepeople.address
 
 import forms.{EmptyForm, InvalidForm}
 import jto.validation.{Path, ValidationError}
@@ -41,7 +41,7 @@ class new_home_date_of_changeSpec extends AmlsSpec with MustMatchers {
         Messages("summary.responsiblepeople") + " - " +
         Messages("title.amls") + " - " + Messages("title.gov")
 
-      def view = views.html.responsiblepeople.new_home_date_of_change(form2, 1, name)
+      def view = views.html.responsiblepeople.address.new_home_date_of_change(form2, 1, name)
 
       doc.getElementsByAttributeValue("class", "link-back") must not be empty
 
@@ -62,7 +62,7 @@ class new_home_date_of_changeSpec extends AmlsSpec with MustMatchers {
           (Path \ "some path") -> Seq(ValidationError("not a message Key"))
         ))
 
-      def view = views.html.responsiblepeople.new_home_date_of_change(form2, 1, name)
+      def view = views.html.responsiblepeople.address.new_home_date_of_change(form2, 1, name)
 
       errorSummary.html() must include("not a message Key")
     }
